# Improve performance flatten 5-20x