--- a/packages/turf-flatten/README.md
+++ b/packages/turf-flatten/README.md
@@ -7,7 +7,7 @@ Flattens any [GeoJSON](GeoJSON) to a [FeatureCollection](http://geojson.org/geoj
 **Parameters**
 
 -   `geojson` **[Feature](http://geojson.org/geojson-spec.html#feature-objects)** any valid [GeoJSON](GeoJSON) with multi-geometry [Feature](http://geojson.org/geojson-spec.html#feature-objects)s
--   `properties` **\[Any]** translate properties to Feature (only applies for [GeometryCollection](http://geojson.org/geojson-spec.html#geometrycollection))
+-   `properties` **\[Any]** translate properties to each [Feature](http://geojson.org/geojson-spec.html#feature-objects) (only applies for [GeometryCollection](http://geojson.org/geojson-spec.html#geometrycollection) & [GeometryObject](GeometryObject)) (optional, default `{}`)
 
 **Examples**