--- a/packages/turf-flatten/index.js
+++ b/packages/turf-flatten/index.js
@@ -1,5 +1,5 @@
 var flatten = require('geojson-flatten');
-var featurecollection = require('@turf/helpers').featureCollection;
+var helpers = require('@turf/helpers');
 
 /**
  * Flattens any {@link GeoJSON} to a {@link FeatureCollection} using [geojson-flatten](https://github.com/mapbox/geojson-flatten).
@@ -21,9 +21,61 @@ var featurecollection = require('@turf/helpers').featureCollection;
  *
  * //=flattened
  */
-
 module.exports = function (geojson) {
+    switch (geojson.type) {
+    case 'MultiPolygon':
+        return flattenMultiPolygon(geojson);
+    case 'MultiLineString':
+        return flattenMultiLineString(geojson);
+    case 'MultiPoint':
+        return flattenMultiPoint(geojson);
+    }
     var flattened = flatten(geojson);
     if (flattened.type === 'FeatureCollection') return flattened;
-    else return featurecollection(flatten(geojson));
+    else return helpers.featureCollection(flatten(geojson));
 };
+
+/**
+ * Flatten MultiPoint
+ *
+ * @private
+ * @param {Feature<MultiPoint>} geojson GeoJSON Feature
+ * @returns {FeatureCollection<Point>} Feature Collection
+ */
+function flattenMultiPoint(geojson) {
+    var points = [];
+    geojson.coordinates.forEach(function (coordinates) {
+        points.push(helpers.point(coordinates));
+    });
+    return helpers.featureCollection(points);
+}
+
+/**
+ * Flatten MultiLineString
+ *
+ * @private
+ * @param {Feature<MultiLineString>} geojson GeoJSON Feature
+ * @returns {FeatureCollection<LineString>} Feature Collection
+ */
+function flattenMultiLineString(geojson) {
+    var lines = [];
+    geojson.coordinates.forEach(function (coordinates) {
+        lines.push(helpers.lineString(coordinates));
+    });
+    return helpers.featureCollection(lines);
+}
+
+/**
+ * Flatten MultiPolygon
+ *
+ * @private
+ * @param {Feature<MultiPolygon>} geojson GeoJSON Feature
+ * @returns {FeatureCollection<Polygon>} Feature Collection
+ */
+function flattenMultiPolygon(geojson) {
+    var polygons = [];
+    geojson.coordinates.forEach(function (coordinates) {
+        polygons.push(helpers.polygon(coordinates));
+    });
+    return helpers.featureCollection(polygons);
+}