--- a/packages/turf-flatten/package.json
+++ b/packages/turf-flatten/package.json
@@ -31,7 +31,10 @@
   "homepage": "https://github.com/Turfjs/turf",
   "devDependencies": {
     "benchmark": "^2.1.3",
-    "tape": "^4.0.0"
+    "geojson-flatten": "^0.2.1",
+    "load-json-file": "^2.0.0",
+    "tape": "^4.0.0",
+    "write-json-file": "^2.0.0"
   },
   "dependencies": {
     "@turf/helpers": "^3.10.5",