--- a/packages/turf-flatten/test/in/GeometryCollection.geojson
+++ b/packages/turf-flatten/test/in/GeometryCollection.geojson
@@ -1,0 +1,25 @@
+{
+	"type": "GeometryCollection",
+	"geometries": [
+		{
+			"type": "Point",
+			"coordinates": [
+				100,
+				0
+			]
+		},
+		{
+			"type": "LineString",
+			"coordinates": [
+				[
+					101,
+					0
+				],
+				[
+					102,
+					1
+				]
+			]
+		}
+	]
+}