--- a/packages/turf-flatten/test/in/MultiLineString.geojson
+++ b/packages/turf-flatten/test/in/MultiLineString.geojson
@@ -1,0 +1,31 @@
+{
+  "type": "Feature",
+  "properties": {
+    "foo": "bar"
+  },
+  "geometry": {
+    "type": "MultiLineString",
+    "coordinates": [
+      [
+        [
+          100,
+          0
+        ],
+        [
+          101,
+          1
+        ]
+      ],
+      [
+        [
+          102,
+          2
+        ],
+        [
+          103,
+          3
+        ]
+      ]
+    ]
+  }
+}