--- a/packages/turf-flatten/test/in/MultiPoint.geojson
+++ b/packages/turf-flatten/test/in/MultiPoint.geojson
@@ -1,0 +1,19 @@
+{
+  "type": "Feature",
+  "properties": {
+    "foo": "bar"
+  },
+  "geometry": {
+    "type": "MultiPoint",
+    "coordinates": [
+      [
+        100,
+        0
+      ],
+      [
+        101,
+        1
+      ]
+    ]
+  }
+}