--- a/packages/turf-flatten/test/out/GeometryCollection.geojson
+++ b/packages/turf-flatten/test/out/GeometryCollection.geojson
@@ -1,0 +1,33 @@
+{
+	"type": "FeatureCollection",
+	"features": [
+		{
+			"type": "Feature",
+			"properties": {},
+			"geometry": {
+				"type": "Point",
+				"coordinates": [
+					100,
+					0
+				]
+			}
+		},
+		{
+			"type": "Feature",
+			"properties": {},
+			"geometry": {
+				"type": "LineString",
+				"coordinates": [
+					[
+						101,
+						0
+					],
+					[
+						102,
+						1
+					]
+				]
+			}
+		}
+	]
+}