--- a/packages/turf-flatten/test/out/MultiLineString.geojson
+++ b/packages/turf-flatten/test/out/MultiLineString.geojson
@@ -1,0 +1,43 @@
+{
+	"type": "FeatureCollection",
+	"features": [
+		{
+			"type": "Feature",
+			"properties": {
+				"foo": "bar"
+			},
+			"geometry": {
+				"type": "LineString",
+				"coordinates": [
+					[
+						100,
+						0
+					],
+					[
+						101,
+						1
+					]
+				]
+			}
+		},
+		{
+			"type": "Feature",
+			"properties": {
+				"foo": "bar"
+			},
+			"geometry": {
+				"type": "LineString",
+				"coordinates": [
+					[
+						102,
+						2
+					],
+					[
+						103,
+						3
+					]
+				]
+			}
+		}
+	]
+}