--- a/packages/turf-flatten/test/out/MultiPoint.geojson
+++ b/packages/turf-flatten/test/out/MultiPoint.geojson
@@ -1,0 +1,31 @@
+{
+	"type": "FeatureCollection",
+	"features": [
+		{
+			"type": "Feature",
+			"properties": {
+				"foo": "bar"
+			},
+			"geometry": {
+				"type": "Point",
+				"coordinates": [
+					100,
+					0
+				]
+			}
+		},
+		{
+			"type": "Feature",
+			"properties": {
+				"foo": "bar"
+			},
+			"geometry": {
+				"type": "Point",
+				"coordinates": [
+					101,
+					1
+				]
+			}
+		}
+	]
+}