--- a/packages/turf-flatten/test/types.ts
+++ b/packages/turf-flatten/test/types.ts
@@ -1,0 +1,4 @@
+import * as flatten from '../'
+
+const multiPolygon: GeoJSON.Feature<GeoJSON.MultiPolygon> = undefined
+flatten(multiPolygon);