--- a/packages/turf-flatten/test/types.ts
+++ b/packages/turf-flatten/test/types.ts
@@ -1,4 +1,9 @@
+import * as path from 'path'
 import * as flatten from '../'
+const load = require('load-json-file')
 
-const multiPolygon: GeoJSON.Feature<GeoJSON.MultiPolygon> = undefined
+const multiPolygon: GeoJSON.Feature<GeoJSON.MultiPolygon> = load.sync(path.join(__dirname, 'in', 'MultiPolygon.geojson'))
 flatten(multiPolygon);
+
+const geometryCollection: GeoJSON.GeometryCollection = load.sync(path.join(__dirname, 'in', 'GeometryCollection.geojson'))
+flatten(geometryCollection);