… packages/turf-flatten/test/in/FeatureCollection.geojson
… packages/turf-flatten/test/in/MultiPolygon.geojson
… packages/turf-flatten/test/out/FeatureCollection.geojson
… packages/turf-flatten/test/out/MultiPolygon.geojson