Add speed improvements for the following geometries:
- FeatureCollection (20x)
- GeometryCollection (10x)
- MultiLineString (7.5x)
- MultiPoint (7.5x)
- MultiPolygon (7.5x)

Also included `properties` parameter.

```javascript
/**
 * @param {*} [properties={}] translate properties to Features (only applies for GeometryCollection & GeometryObject)
```

**Before**
```
FeatureCollection x 257,481 ops/sec ±0.60% (90 runs sampled)
GeometryCollection x 258,479 ops/sec ±0.90% (90 runs sampled)
MultiLineString x 464,694 ops/sec ±1.07% (90 runs sampled)
MultiPoint x 462,874 ops/sec ±0.87% (88 runs sampled)
MultiPolygon x 459,886 ops/sec ±1.06% (89 runs sampled)
```

**After**
```
FeatureCollection x 4,804,435 ops/sec ±0.96% (88 runs sampled)
GeometryCollection x 2,661,693 ops/sec ±1.41% (85 runs sampled)
MultiLineString x 3,913,849 ops/sec ±2.71% (86 runs sampled)
MultiPoint x 3,259,684 ops/sec ±1.06% (87 runs sampled)
MultiPolygon x 3,473,415 ops/sec ±1.32% (87 runs sampled)
```

CC: #595