@rowanwins 2017-03-07
@DenisCarriere - Any reason why you didn't make these changes upstream in geojson-flatten repo?