@DenisCarriere 2017-03-07
It was an entire re-write... I could of.

Most of the `geojson-<module>` should be merged or maintained in TurfJS and be "Deprecated", no reason to maintain two libraries with sub-dependencies everywhere.

TurfJS handles our linting, browser build, typescript definitions, benchmarks, better to keep everything in one place.